PCLCommon.jl

PCLCommon — Module.

PCL common types, functions and utilities. The primary export is the PointCloud{PointT} (aliased to PointCloudPtr{PointT}), which represents a shared pointer of a point cloud (i.e. pcl::PointCloud<PointT>::Ptr) in PCL. You can create point clouds as follows:

using PCLCommon

# Create empty point cloud
cloud = PointCloud{PointXYZRGBA}()

using PCLCommon
using PCLIO

# Create and load point cloud from a PCD file
cloud = PointCloud{PointXYZRGB}("your_pcd_file.pcd")

source

Index

PCLCommon.BoostSharedPtr

PCLCommon.PointCloud

PCLCommon.deg2rad

PCLCommon.use_count

PCLCommon.Axis

PCLCommon.BRISKSignature512

PCLCommon.Boundary

PCLCommon.CPPFSignature

PCLCommon.Correspondence

PCLCommon.CorrespondencePtr

PCLCommon.CorrespondenceVal

PCLCommon.Correspondences

PCLCommon.CorrespondencesPtr

PCLCommon.CorrespondencesVal

PCLCommon.ESFSignature640

PCLCommon.FPFHSignature33

PCLCommon.GRSDSignature21

PCLCommon.IntensityGradient

PCLCommon.InterestPoint

PCLCommon.Label

PCLCommon.ModelCoefficients

PCLCommon.ModelCoefficientsPtr

PCLCommon.ModelCoefficientsVal

PCLCommon.MomentInvariants

PCLCommon.Narf36

PCLCommon.Normal

PCLCommon.NormalBasedSignature12

PCLCommon.PCLBase

PCLCommon.PCLPointCloud2

PCLCommon.PCLPointCloud2Ptr

PCLCommon.PCLPointCloud2Val

PCLCommon.PFHRGBSignature250

PCLCommon.PFHSignature125

PCLCommon.PPFRGBSignature

PCLCommon.PPFSignature

PCLCommon.PointCloudPtr

PCLCommon.PointCloudVal

PCLCommon.PointDEM

PCLCommon.PointIndices

PCLCommon.PointIndicesPtr

PCLCommon.PointIndicesVal

PCLCommon.PointNormal

PCLCommon.PointSurfel

PCLCommon.PointUV

PCLCommon.PointWithRange

PCLCommon.PointWithScale

PCLCommon.PointWithViewpoint

PCLCommon.PointXY

PCLCommon.PointXYZ

PCLCommon.PointXYZHSV

PCLCommon.PointXYZI

PCLCommon.PointXYZINormal

PCLCommon.PointXYZL

PCLCommon.PointXYZLNormal

PCLCommon.PointXYZRGB

PCLCommon.PointXYZRGBA

PCLCommon.PointXYZRGBL

PCLCommon.PointXYZRGBNormal

PCLCommon.PrincipalCurvatures

PCLCommon.PrincipalRadiiRSD

PCLCommon.RangeImage

PCLCommon.RangeImagePtr

PCLCommon.RangeImageVal

PCLCommon.ReferenceFrame

PCLCommon.SHOT1344

PCLCommon.SHOT352

PCLCommon.ShapeContext1980

PCLCommon.UniqueShapeContext1960

PCLCommon.VFHSignature308

PCLCommon.@boostsharedptr

PCLCommon.@defconstructor

PCLCommon.@defpcltype

PCLCommon.@defptrconstructor

Reference

PCLCommon.BoostSharedPtr — Constant.

boost::shared_ptr<T>

source

PCLCommon.PointCloud — Constant.

Pointer representation for pcl::PointCloud<PointT> in C++

typealias of PointCloudPtr

source

PCLCommon.use_count — Method.

use_count(s)

Returns reference count

source

PCLCommon.Axis — Type.

pcl::Axis

source

PCLCommon.BRISKSignature512 — Type.

pcl::BRISKSignature512

source

PCLCommon.Boundary — Type.

pcl::Boundary

source

PCLCommon.CPPFSignature — Type.

pcl::CPPFSignature

source

PCLCommon.Correspondence — Type.

Pointer representation for pcl::Correspondence in C++

typealias of CorrespondencePtr

source

PCLCommon.Correspondences — Type.

Pointer representation for pcl::Correspondences in C++

typealias of CorrespondencesPtr

source

PCLCommon.ESFSignature640 — Type.

pcl::ESFSignature640

source

PCLCommon.FPFHSignature33 — Type.

pcl::FPFHSignature33

source

PCLCommon.GRSDSignature21 — Type.

pcl::GRSDSignature21

source

PCLCommon.IntensityGradient — Type.

pcl::IntensityGradient

source

PCLCommon.InterestPoint — Type.

pcl::InterestPoint

source

PCLCommon.Label — Type.

pcl::Label

source

PCLCommon.ModelCoefficients — Type.

Pointer representation for pcl::ModelCoefficients in C++

typealias of ModelCoefficientsPtr

source

PCLCommon.MomentInvariants — Type.

pcl::MomentInvariants

source

PCLCommon.Narf36 — Type.

pcl::Narf36

source

PCLCommon.Normal — Type.

pcl::Normal

source

PCLCommon.NormalBasedSignature12 — Type.

pcl::NormalBasedSignature12

source

PCLCommon.PCLBase — Type.

Similar to pcl::PCLBase, for dispatch

source

PCLCommon.PCLPointCloud2 — Type.

Pointer representation for pcl::PCLPointCloud2 in C++

typealias of PCLPointCloud2Ptr

source

PCLCommon.PFHRGBSignature250 — Type.

pcl::PFHRGBSignature250

source

PCLCommon.PFHSignature125 — Type.

pcl::PFHSignature125

source

PCLCommon.PPFRGBSignature — Type.

pcl::PPFRGBSignature

source

PCLCommon.PPFSignature — Type.

pcl::PPFSignature

source

PCLCommon.PointDEM — Type.

pcl::PointDEM

source

PCLCommon.PointIndices — Type.

Pointer representation for pcl::PointIndices in C++

typealias of PointIndicesPtr

source

PCLCommon.PointNormal — Type.

pcl::PointNormal

source

PCLCommon.PointSurfel — Type.

pcl::PointSurfel

source

PCLCommon.PointUV — Type.

pcl::PointUV

source

PCLCommon.PointWithRange — Type.

pcl::PointWithRange

source

PCLCommon.PointWithScale — Type.

pcl::PointWithScale

source

PCLCommon.PointWithViewpoint — Type.

pcl::PointWithViewpoint

source

PCLCommon.PointXY — Type.

pcl::PointXY

source

PCLCommon.PointXYZ — Type.

pcl::PointXYZ

source

PCLCommon.PointXYZHSV — Type.

pcl::PointXYZHSV

source

PCLCommon.PointXYZI — Type.

pcl::PointXYZI

source

PCLCommon.PointXYZINormal — Type.

pcl::PointXYZINormal

source

PCLCommon.PointXYZL — Type.

pcl::PointXYZL

source

PCLCommon.PointXYZLNormal — Type.

pcl::PointXYZLNormal

source

PCLCommon.PointXYZRGB — Type.

pcl::PointXYZRGB

source

PCLCommon.PointXYZRGBA — Type.

pcl::PointXYZRGBA

source

PCLCommon.PointXYZRGBL — Type.

pcl::PointXYZRGBL

source

PCLCommon.PointXYZRGBNormal — Type.

pcl::PointXYZRGBNormal

source

PCLCommon.PrincipalCurvatures — Type.

pcl::PrincipalCurvatures

source

PCLCommon.PrincipalRadiiRSD — Type.

pcl::PrincipalRadiiRSD

source

PCLCommon.RangeImage — Type.

Pointer representation for pcl::RangeImage in C++

typealias of RangeImagePtr

source

PCLCommon.ReferenceFrame — Type.

pcl::ReferenceFrame

source

PCLCommon.SHOT1344 — Type.

pcl::SHOT1344

source

PCLCommon.SHOT352 — Type.

pcl::SHOT352

source

PCLCommon.ShapeContext1980 — Type.

pcl::ShapeContext1980

source

PCLCommon.UniqueShapeContext1960 — Type.

pcl::UniqueShapeContext1960

source

PCLCommon.VFHSignature308 — Type.

pcl::VFHSignature308

source

PCLCommon.@boostsharedptr — Macro.

@boostsharedptr(name, args)

creates boost::shared_ptr expression (for package development)

Parameters

name : C++ type name

args : argments will be passed to constructor

template parameters in 1st argment and embedding expression in 2nd argment must be escaped.

Returns

handle : C++ boost shared pointer of a given C++ type name and argments

Example

handle = @boostsharedptr(
    "pcl::visualization::PointCloudColorHandlerRGBField<\$T>",
    "\$(cloud.handle)")

source

PCLCommon.@defconstructor — Macro.

@defconstructor(expr, cxxname)

Defines convenient constructor for value types

Parameters

expr : Constructor expression

cxxname : C++ type name

Examples

@defptrconstructor PointCloudVal{PointT}() "pcl::PointCloud"

source

PCLCommon.@defpcltype — Macro.

@defpcltype(expr, cxxname)

A macro to define PCL convenient types. Mostly intended to be used by package development.

Parameters

expr : Julia type name (can have type params)

cxxname : C++ type name

Example

@defpcltype PointCloud{T} "pcl::PointCloud"

which defines Julia wrapper types for C++ types:

PointCloudPtr{T}: boost shared pointer of pcl::PointCloud<T> (i.e. pcl::PointCloud<T>::Ptr) wrapper

PointCloudVal{T}: pcl::PointCloud<T> value wrapper

PointCloud{T}: type aliased to PointCloudPtr{T}

and also Cxx types:

pclPointCloudPtr{T}: pcl::PointCloud<T>::Ptr

pclPointCloudVal{T}: pcl::PointCloud<T>

With the combination of @defptrconstructor, you can then use pcl::PointCloud<T>::Ptr (entirely used in PCL tutorials) as follows:

cloud = PointCloud{PointXYZ}()

Note that PointCloud{T} is a Julia wrapper, you can get the C++ representation as PCLCommon.handle(cloud) or use handle = pclPointCloud{PointXYZ}() which returns C++ type.

source

PCLCommon.@defptrconstructor — Macro.

@defptrconstructor(expr, cxxname)

Defines convenient constructor for point types

Parameters

expr : Constructor expression

cxxname : C++ type name

Examples

@defptrconstructor PointCloud{PointT}() "pcl::PointCloud"

source

Base.convert — Method.

convert(?, cloud)

Converts a point cloud to a different type of point cloud

Examples

cloud = PointCloud{PointXYZRGB}()
xyz_cloud = convert(PointCloud{PointXYZ}, cloud)

source

PCLCommon.deg2rad — Method.

pcl::deg2rad

source

PCLCommon.CorrespondencePtr — Type.

Pointer representation for pcl::Correspondence in C++

source

PCLCommon.CorrespondenceVal — Type.

Value representation for pcl::Correspondence in C++

source

PCLCommon.CorrespondencesPtr — Type.

Pointer representation for pcl::Correspondences in C++

source

PCLCommon.CorrespondencesVal — Type.

Value representation for pcl::Correspondences in C++

source

PCLCommon.ModelCoefficientsPtr — Type.

Pointer representation for pcl::ModelCoefficients in C++

source

PCLCommon.ModelCoefficientsVal — Type.

Value representation for pcl::ModelCoefficients in C++

source

PCLCommon.PCLPointCloud2Ptr — Type.

Pointer representation for pcl::PCLPointCloud2 in C++

source

PCLCommon.PCLPointCloud2Val — Type.

Value representation for pcl::PCLPointCloud2 in C++

source

PCLCommon.PointCloudPtr — Type.

Pointer representation for pcl::PointCloud<PointT> in C++

source

PCLCommon.PointCloudVal — Type.

Value representation for pcl::PointCloud<PointT> in C++

source

PCLCommon.PointIndicesPtr — Type.

Pointer representation for pcl::PointIndices in C++

source

PCLCommon.PointIndicesVal — Type.

Value representation for pcl::PointIndices in C++

source

PCLCommon.RangeImagePtr — Type.

Pointer representation for pcl::RangeImage in C++

source

PCLCommon.RangeImageVal — Type.

Value representation for pcl::RangeImage in C++

source