LibPCL.jl

LibPCL — Module.

A special package that manages PCL binary depedencies

Pkg.build("LibPCL")

try to search your system PCL libraries and its dependencies, and throws errors if any issues. If PCL librarires are not found, it will install fresh PCL libraries into deps directrory, but not recommended unless if you have perfect requiremsnts to build PCL, since it's a little hard to build.

source

Index

LibPCL.find_library_e

Reference

LibPCL.find_library_e — Function.

find_library_e(mod)
find_library_e(mod, libdirs)
find_library_e(mod, libdirs, ext)

It tries to search the specified library by name. Not exported, but meant to be used by other PCL packages.

Parameters

mod : Module name

libdirs : library seach directries (default is dir of libpclcommon)

ext : library extention name (e.g. .so)

Retures

libpath : library path if found, othrewise return C_NULL

Examples

From the PCLVisualization package,

const libpcl_visualization = LibPCL.find_library_e("libpcl_visualization")
try
    Libdl.dlopen(libpcl_visualization, Libdl.RTLD_GLOBAL)
catch e
    warn("You might need to set DYLD_LIBRARY_PATH to load dependencies proeprty.")
    rethrow(e)
end

source